+++ packages/sdoc-editor/src/extension/commons/insert-element-dialog/index.js
@@ -11,7 +11,7 @@ import FormulaModal from '../../plugins/formula/menu/formula-modal.js';
 import { generateImageInfos, insertImage } from '../../plugins/image/helpers';
 import AddLinkDialog from '../../plugins/link/dialog/add-link-dialog';
 import { CustomTableSizeDialog, SplitCellSettingDialog } from '../../plugins/table/dialogs';
-import { VIDEO_MAX_SIZE_5MB } from '../../plugins/video/constants/index.js';
+import { formatSeadocVideoSizeLimit, SEADOC_VIDEO_SIZE_LIMIT } from '../../plugins/video/constants/index.js';
 import AddVideoLinkDialog from '../../plugins/video/dialog/add-video-link-dialog/index.js';
 import { insertVideo } from '../../plugins/video/helpers';
 import FileLinkInsertDialog from '../file-insert-dialog/index.js';
@@ -49,15 +49,15 @@ const InsertElementDialog = ({ editor }) => {
 
   const handleDisplayAlert = useCallback(() => {
     setTimeout(() => {
-      toaster.warning(`${t('The_current_version_does_not_support_>5MB_video_file')}`, { duration: 3 });
+      toaster.warning(t('The_video_file_exceeds_the_size_limit', { size: formatSeadocVideoSizeLimit() }), { duration: 3 });
     }, 0);
-  // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
+  }, [t]);
 
   const onVideoFileChanged = useCallback((event) => {
     const files = event.target.files;
-    // Show warning for 3s and no further insertion if video file is more than 5MB
-    if (files[0].size > VIDEO_MAX_SIZE_5MB) {
+    const videoSizeLimit = SEADOC_VIDEO_SIZE_LIMIT;
+    // Show warning for 3s and no further insertion if the video file exceeds the default limit.
+    if (files[0].size > videoSizeLimit) {
       handleDisplayAlert();
       event.target.value = null;
       return;